Consolidating get() and get(timeout) functions.
These two can be implemented in a single function. I
merged them because this is another point where we
have the information that "certain" connection is
gone. The other is in the CoreStream when we are
posting I/O requests.
May be in future, we want to take some steps here.